Moved page tests from subfolder into main folder.
