+++ app/pages/GenerateRandomStringPage.test.js
@@ -3,17 +3,16 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import { dummyTranslatorProps } from 'config/tests';
+import { dummyRouterProps } from 'config/tests';
 
-import { PureLibraryPage } from '../LibraryPage';
+import { PureGenerateRandomStringPage } from './GenerateRandomStringPage';
 
-describe(`LibraryPage`, (): void => {
+describe(`GenerateRandomStringPage`, (): void => {
 
   it(`renders without errors`, (): void => {
     const enzymeWrapper = shallow(
-      <PureLibraryPage
-        {...dummyTranslatorProps}
-        currentUserId="dummyUserId"
+      <PureGenerateRandomStringPage
+        {...dummyRouterProps}
       />,
     );
     expect(enzymeWrapper.isEmptyRender()).toEqual(false);
