+++ app/pages/ProfilePage.test.js
@@ -5,15 +5,24 @@ import { shallow } from 'enzyme';
 
 import { dummyTranslatorProps, dummyRouterProps } from 'config/tests';
 
-import { PureEditorPage } from '../EditorPage';
+import { PureProfilePage } from './ProfilePage';
 
-describe(`EditorPage`, (): void => {
+describe(`ProfilePage`, (): void => {
 
   it(`renders without errors`, (): void => {
+    const dummyAccount = {
+      id: 'johanjohan',
+      email: 'john.doe@example.com',
+      firstName: 'John',
+      lastName: 'Doe',
+    };
+
     const enzymeWrapper = shallow(
-      <PureEditorPage
+      <PureProfilePage
         {...dummyTranslatorProps}
         {...dummyRouterProps}
+        account={dummyAccount}
+        userId="abcdefghij"
       />,
     );
     expect(enzymeWrapper.isEmptyRender()).toEqual(false);
